+++ testProject/src/main/kotlin/foo/bar/baz/KotlinNestedClasses.kt
@@ -1,4 +1,4 @@
-// FileTestVector{baseFqClassNames: foo.bar.baz.KotlinNestedClasses, containingFqClassNames: foo.bar.baz.KotlinNestedClasses|foo.bar.baz.KotlinNestedClasses$methodWithLocalClass$1|foo.bar.baz.KotlinNestedClasses$Inner|foo.bar.baz.KotlinNestedClasses$Nested|foo.bar.baz.KotlinNestedClasses$Nested$Companion|foo.bar.baz.KotlinNestedClasses$Companion }
+// FileTestVector{baseFqClassNames: foo.bar.baz.KotlinNestedClasses, containingFqClassNames: foo.bar.baz.KotlinNestedClasses|foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$1|foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$LocalClass|foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$myObject$1|foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$myObject$1|foo.bar.baz.KotlinNestedClasses$Inner|foo.bar.baz.KotlinNestedClasses$Nested|foo.bar.baz.KotlinNestedClasses$Nested$Companion|foo.bar.baz.KotlinNestedClasses$Companion }
 package foo.bar.baz;
 
 class KotlinNestedClasses {
@@ -7,10 +7,22 @@ class KotlinNestedClasses {
     // PsiElementTestVector{reference: METHOD|KotlinNestedClasses#method, baseFqClassName: foo.bar.baz.KotlinNestedClasses, expectedFqClassNames: foo.bar.baz.KotlinNestedClasses}
   }
 
-  fun methodWithLocalClass() {
+  fun `method with local and anonymous class`() {
+    val myObject = object {
+      fun methodInObjectWithVariable() {
+        // PsiElementTestVector{reference: METHOD|KotlinNestedClasses#methodInObjectWithVariable, baseFqClassName: foo.bar.baz.KotlinNestedClasses, expectedFqClassNames: foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$myObject$1, sourceFileOnly: true}
+      }
+    }
+
+    class LocalClass {
+      fun methodInLocalClass() {
+        // PsiElementTestVector{reference: METHOD|KotlinNestedClasses#methodInLocalClass, baseFqClassName: foo.bar.baz.KotlinNestedClasses, expectedFqClassNames: foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$LocalClass, sourceFileOnly: true}
+      }
+    }
+
     object: Runnable {
       override fun run() {
-        // PsiElementTestVector{reference: METHOD|KotlinNestedClasses#run, baseFqClassName: foo.bar.baz.KotlinNestedClasses, expectedFqClassNames: foo.bar.baz.KotlinNestedClasses, sourceFileOnly: true}
+        // PsiElementTestVector{reference: METHOD|KotlinNestedClasses#run, baseFqClassName: foo.bar.baz.KotlinNestedClasses, expectedFqClassNames: foo.bar.baz.KotlinNestedClasses$method with local and anonymous class$1, sourceFileOnly: true}
       }
     }
   }